dom/nodes/Element-lastElementChild-svg.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="100%" height="100%" version="1.1" viewBox="0 0 400 400"><title>lastElementChild</title><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><text x="200" y="40" fill="#000" font-size="25" text-anchor="middle">Test of lastElementChild</text><text id="parentEl" x="200" y="70" fill="#000" font-size="20" text-anchor="middle">The result of <tspan id="first_element_child">this test</tspan> is <tspan id="last_element_child" font-weight="bold">not</tspan> known.</text><h:script><![CDATA[
test(function() {
  var parentEl = document.getElementById("parentEl");
  var lec = parentEl.lastElementChild;
  assert_true(!!lec)
  assert_equals(lec.nodeType, 1)
  assert_equals(lec.getAttribute("id"), "last_element_child")
})
]]></h:script></svg>
